move batches component und util into nextjs/components folder
